splittting gui  into file tree.
